requirements.txt
apache-airflow
snowflake-connector-python
fredapi
pandas
mlflow
scikit-learn
prophet
python-dotenv
streamlit
plotly
pandas
xgboost
joblib